+++ fs/test.ts
@@ -1,0 +1,12 @@
+import "./path/test.ts";
+import "./walk_test.ts";
+import "./globrex_test.ts";
+import "./glob_test.ts";
+import "./exists_test.ts";
+import "./empty_dir_test.ts";
+import "./ensure_dir_test.ts";
+import "./ensure_file_test.ts";
+import "./move_test.ts";
+import "./read_json_test.ts";
+import "./write_json_test.ts";
+import "./utils_test.ts";
