+++ http/test.ts
@@ -1,0 +1,3 @@
+
+import "./server_test.ts";
+import "./file_server_test.ts";
