+++ io/test.ts
@@ -1,3 +1,5 @@
 import "./bufio_test.ts";
 import "./ioutil_test.ts";
 import "./util_test.ts";
+import "./writers_test.ts";
+import "./readers_test.ts";
