+++ prettier/test.ts
@@ -1,0 +1,1 @@
+import "./main_test.ts";
