+++ strings/test.ts
@@ -1,0 +1,1 @@
+import "./pad_test.ts";
