+++ test.ts
@@ -4,32 +4,15 @@ import "./colors/test.ts";
 import "./datetime/test.ts";
 import "./examples/test.ts";
 import "./flags/test.ts";
-import "./io/bufio_test.ts";
-import "./io/ioutil_test.ts";
-import "./io/util_test.ts";
-import "./io/writers_test.ts";
-import "./io/readers_test.ts";
-import "./fs/path/test.ts";
-import "./fs/walk_test.ts";
-import "./fs/globrex_test.ts";
-import "./fs/glob_test.ts";
-import "./fs/exists_test.ts";
-import "./fs/empty_dir_test.ts";
-import "./fs/ensure_dir_test.ts";
-import "./fs/ensure_file_test.ts";
-import "./fs/move_test.ts";
-import "./fs/read_json_test.ts";
-import "./fs/write_json_test.ts";
-import "./fs/utils_test.ts";
 import "./io/test.ts";
-import "./http/server_test.ts";
-import "./http/file_server_test.ts";
+import "./fs/test.ts";
+import "./io/test.ts";
+import "./http/test.ts";
 import "./log/test.ts";
 import "./media_types/test.ts";
-import "./multipart/formfile_test.ts";
-import "./multipart/multipart_test.ts";
-import "./prettier/main_test.ts";
-import "./strings/pad_test.ts";
+import "./multipart/test.ts";
+import "./prettier/test.ts";
+import "./strings/test.ts";
 import "./testing/test.ts";
 import "./textproto/test.ts";
 import "./ws/test.ts";
